Organize

+++ src/BackfireSoundGenerator.cpp
@@ -11,16 +11,17 @@
 float BackfireSoundGenerator::getAmplitude() const { return amplitude; }
 void BackfireSoundGenerator::setAmplitude(float g) { amplitude = std::max(0.0f, g); }
 
-BackfireSoundGenerator::BackfireSoundGenerator(float sampleRate) : sampleRate(sampleRate), amplitude(0.6f), intensity(0.0f) { srand(static_cast<unsigned>(time(nullptr))); }
+BackfireSoundGenerator::BackfireSoundGenerator(float sampleRate) : sampleRate(sampleRate), amplitude(0.6f), intensity(0.0f) {
+    srand(static_cast<unsigned>(time(nullptr)));
+}
 
 void BackfireSoundGenerator::setIntensity(float newIntensity) { intensity = std::clamp(newIntensity, 0.0f, 1.0f); }
 
 float BackfireSoundGenerator::getIntensity() const { return intensity; }
 
-
 void BackfireSoundGenerator::update() {
     // Random pop trigger based on intensity
-    if (popQueue.empty() && ((float)rand() / RAND_MAX) < intensity * 0.015f) {
+    if (popQueue.empty() && ((float)rand() / (float) RAND_MAX) < intensity * 0.015f) {
         triggerPopBurst();
     }
 
@@ -36,7 +37,7 @@ void BackfireSoundGenerator::update() {
 float BackfireSoundGenerator::getSample() {
     float sample = 0.0f;
     for (auto &pop : popQueue) {
-        float noise = ((float)rand() / RAND_MAX) * 2.0f - 1.0f;
+        float noise = ((float)rand() / (float) RAND_MAX) * 2.0f - 1.0f;
         float freq = pop.freq;
         sample += pop.envelope * bandpass(noise, freq, 0.9f);
     }
